feat(expo): Rename package to `@clerk/expo`
