+++ .changeset/strong-bars-learn.md
@@ -1,7 +1,7 @@
 ---
 '@clerk/chrome-extension': patch
 '@clerk/elements': patch
-'@clerk/clerk-expo': patch
+'@clerk/expo': patch
 '@clerk/nextjs': patch
 '@clerk/react-router': patch
 '@clerk/tanstack-react-start': patch
