+++ .changeset/tricky-humans-stand.md
@@ -1,0 +1,5 @@
+---
+'@clerk/expo': major
+---
+
+Rename package to `@clerk/expo`.
