+++ packages/expo/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@clerk/clerk-expo",
+  "name": "@clerk/expo",
   "version": "2.16.1",
   "description": "Clerk React Native/Expo library",
   "keywords": [
