+++ packages/expo/src/index.ts
@@ -18,7 +18,7 @@ export * from './provider/ClerkProvider';
 export * from './hooks';
 export * from './components';
 
-// Override Clerk React error thrower to show that errors come from @clerk/clerk-expo
+// Override Clerk React error thrower to show that errors come from @clerk/expo
 setErrorThrowerOptions({ packageName: PACKAGE_NAME });
 
 export type { TokenCache } from './cache/types';
