+++ packages/expo/src/secure-store/index.ts
@@ -1,7 +1,7 @@
 import { resourceCache } from '../resource-cache';
 
 /**
- * @deprecated Use `resourceCache` from `@clerk/clerk-expo/resource-cache` instead.
+ * @deprecated Use `resourceCache` from `@clerk/expo/resource-cache` instead.
  */
 const secureStore = resourceCache;
