+++ packages/expo/vitest.setup.mts
@@ -1,6 +1,6 @@
 import { beforeAll } from 'vitest';
 
-globalThis.PACKAGE_NAME = '@clerk/clerk-expo';
+globalThis.PACKAGE_NAME = '@clerk/expo';
 globalThis.PACKAGE_VERSION = '0.0.0-test';
 
 beforeAll(() => {});
