+++ packages/upgrade/src/guide-generators/core-2/expo/index.js
@@ -13,7 +13,7 @@ const version = 'core-2';
 const semverVersion = 'v1';
 const name = 'expo';
 const properName = 'Expo';
-const packageName = '@clerk/clerk-expo';
+const packageName = '@clerk/expo';
 const cwd = `${version}/${name}`;
 
 async function generate() {
