+++ scripts/snapshot.mjs
@@ -17,7 +17,7 @@ const snapshot = `---
 '@clerk/react': patch
 '@clerk/remix': patch
 '@clerk/types': patch
-'@clerk/clerk-expo': patch
+'@clerk/expo': patch
 '@clerk/express': patch
 '@clerk/testing': patch
 '@clerk/elements': patch
