+++ turbo.json
@@ -284,7 +284,7 @@
       "outputLogs": "new-only"
     },
     "//#test:integration:expo-web": {
-      "dependsOn": ["@clerk/testing#build", "@clerk/clerk-js#build", "@clerk/clerk-expo#build", "@clerk/react#build"],
+      "dependsOn": ["@clerk/testing#build", "@clerk/clerk-js#build", "@clerk/expo#build", "@clerk/react#build"],
       "env": ["CLEANUP", "DEBUG", "E2E_*", "INTEGRATION_INSTANCE_KEYS"],
       "inputs": ["integration/**"],
       "outputLogs": "new-only"
